Fix compilation bugs
**Multiple definitions of `ui_alert` / `ui_alert_id` / `ui_custom_alert`** --- `ui_screens.h` had been declaring these functions (defined in `ui.c`) as part of its public interface. Under LTO, the compiler saw two translation units claiming to define them. Fixed by rewriting `ui_screens.h` to only declare what `ui_screens.c` actually provides: the `ScreenID` enum, `ui_goto_screen`, `ui_game_loop`, and `palm_form_to_screen`.

**`GEN_FrmAlert` / `GEN_FrmCustomAlert` undefined** --- `palmcompat.h` had `extern` declarations for these but nothing defined them. Added `__attribute__((used))` wrapper functions in `ui.c` that delegate to `ui_alert_id` / `ui_custom_alert`. These exist because some game files compiled without palmcompat.h macros active emit direct calls to `GEN_FrmAlert`.

**`FindSystem` undefined** --- was declared with a local `extern` inside `GEN_FrmDoDialog`, which LTO can lose. Moved to file-scope `extern char FindSystem[]` at the top of `ui.c`.

**`DeterminePrices` / `StandardPrice` / `NextSystemWithinRange` / `DetermineShipPrices` / `GetForHire` / `TotalShields` / `TotalShieldStrength`** --- all were declared with inline `extern` inside functions in `ui_screens.c`. LTO optimises those away when it doesn't see a cross-unit call site. Moved all 7 to file-scope extern declarations at the top of `ui_screens.c`.

**`ui_msgbox` undefined** --- leftover call from the pre-rewrite era. All calls were already replaced with `ui_alert` / `ui_confirm` in previous sessions; this was a comment remnant in `palmcompat.h`, harmless.

**`main()` return type mismatch** --- changed `u16 main(void)` to `int main(void)` to match SGDK's expected signature.

+++ Src/main.c
@@ -37,7 +37,7 @@ static void show_splash(void)
 /* -----------------------------------------------------------------------
  * Main entry point
  * --------------------------------------------------------------------- */
-u16 main(void)
+int main(void)
 {
     VDP_setScreenWidth320();
     kprintf("main: start");
